Refactored gateways into interfaces and changed controllers to gateways and vice versa. Contributors: Albert: aj.luo@mail.utoronto.ca Yasmin: yasamin.nourijelyani@mail.utoronto.ca Umayrah: umayrah.chonee@mail.utoronto.ca
