+++ src/Main/OptionsPackage/SearchController.java
@@ -1,48 +1,61 @@
 package OptionsPackage;
 
+import InputAndOutput.SystemInOut;
 import ProductFunctions.Product;
-import ProductFunctions.ProductReadWriter;
-import UserFunctions.UserReadWriter;
+import UserFunctions.User;
+import UserFunctions.CartManager;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashMap;
-
-/**
- * take user input from command line interface and
- * Takes in user input tag words and and returns a list of all products that match that description to the user,
- * taking this information from a the masterManager which accese the tageProduct dictionary in master.
- * That dictionary has keys as the tag words for each product, and a list of product associated with that tag
- * word.
- */
+import java.util.List;
 
 public class SearchController {
 
+    public List<Product> allowSearch(SystemInOut input) throws IOException, ClassNotFoundException {
+        SearchGateway searchController = new SearchGateway();
 
-    public ArrayList<Product> searchProducts(String tag) throws IOException, ClassNotFoundException {
-
-        File file = new File("src/Main/product.ser");
-        if (file.length() == 0){
-            UserReadWriter rw = new UserReadWriter();
-            HashMap<String, Object> emptyHashMap = new HashMap<>();
-            rw.saveToFile("src/Main/product.ser", emptyHashMap);
-        }
-
-        ProductReadWriter rw = new ProductReadWriter();
-        HashMap<String, Object> productSavedDict = rw.readFromFile("src/Main/product.ser");
-
-        if (productSavedDict.containsKey(tag)){
-            return (ArrayList<Product>) productSavedDict.get(tag);
-        }
-        // if the user does not exist, return a user with an empty username, which the empty username is unaccepted
-        // username anyways
-
-
-        // an empty list since the tag does not exist
-        return new ArrayList<>();
-
+        input.sendOutput("What is a tag word for your product of interest?");
+        String tagOfInterest = input.getInput();
+        return searchController.searchProducts(tagOfInterest);
 
     }
 
+    public void allowBuy(SystemInOut input, User user) throws IOException, ClassNotFoundException {
+
+        TagInterestItemsPresenter presenter = new TagInterestItemsPresenter();
+        List<Product> productsOfInterest = this.allowSearch(input);
+        presenter.presentTagList(productsOfInterest, input); // check if product list is empty
+
+        input.sendOutput("Would you like to purchase one of the items?" +
+                "enter the number of your choice\n 1.Yes\n2.No");
+        String decisionToBuy = input.getInput();
+
+        if (decisionToBuy.equals("1")){
+            // then the user wants to buy
+            input.sendOutput("Please select the index of the item that you want to buy. The index is the " +
+                    "integer value for the position of the item on the list, with the first item being at " +
+                    "the 0th index.");
+            String itemIndex = input.getInput();
+            int indexInt = Integer.parseInt(itemIndex);
+            if (indexInt < productsOfInterest.size()){
+                Product itemOfInterest = productsOfInterest.get(indexInt);
+                // add this product to the cart
+                CartManager cart = new CartManager();
+                cart.addToCart(itemOfInterest, user);
+
+            }else{
+                input.sendOutput("incorrect index, try again");
+                this.allowBuy(input, user); // let user input new index instead of searching again
+            }
+
+
+        }
+        else {
+            // SearchController search = new SearchController();
+            // search.allowSearch(input);
+            // should ask if user wants to exit out of search here or search again
+            allowBuy(input, user);
+        }
+
     }
+    }
+
