+++ src/Main/Run.java
@@ -1,5 +1,5 @@
 import InputAndOutput.SystemInOut;
-import login.WelcomePageGateway;
+import login.WelcomePageController;
 
 import java.io.IOException;
 
@@ -7,7 +7,7 @@ public class Run {
 
     public static void main(String[] args) throws IOException {
         SystemInOut inOut = new SystemInOut();
-        login.WelcomePageGateway initialLogin = new login.WelcomePageGateway();
+        WelcomePageController initialLogin = new WelcomePageController();
         initialLogin.userLoginDecision(inOut);
 
         }
