+++ src/Main/UserFunctions/CartManager.java
@@ -12,7 +12,6 @@ public class CartManager {
     public void addToCart(Product item, User user){
         List<Product> userCart = user.getShoppingCart();
         userCart.add(item);
-
     }
