+++ src/Main/login/SignInController.java
@@ -1,41 +1,28 @@
 package login;
 
 import InputAndOutput.SystemInOut;
-import UserFunctions.ReadWriter;
 import UserFunctions.User;
-import UserFunctions.UserReadWriter;
 
 import java.io.IOException;
-import java.util.HashMap;
 
 public class SignInController {
 
+    String getUsername(SystemInOut inOut) throws IOException {
+        inOut.sendOutput("What is your username? (please do not leave this field empty)");
+        String username = inOut.getInput();
+        while (username.equals("")){
+            inOut.sendOutput("Invalid Username, please reenter your username.");
+            this.getUsername(inOut);
+        }return username;
 
-
-    /**
-     * Return a user, either a new one with the given username, or an old user, that was previously saved in our system
-     * @param username The username inputted from the user
-     * @param inOut To send output to the user
-     * @return A newly created or old user
-     * @throws IOException
-     * @throws ClassNotFoundException
-     */
-
-    public User allowSignIn(String username, SystemInOut inOut) throws IOException, ClassNotFoundException {
-
-        // access the serialized file for this user.
-        UserReadWriter rw = new UserReadWriter();
-        HashMap<String, Object> usersSavedDict = rw.readFromFile("src/Main/user.ser");
-
-        if (usersSavedDict.containsKey(username)){
-            return (User)usersSavedDict.get(username);
         }
-        // if the user does not exist, return a user with an empty username, which the empty username is unaccepted
-        // username anyways
 
-        return new User("");
+        }
 
 
-    }
 
-}
+
+
+
+
+
