+++ src/Main/login/SignInGatewayInterface.java
@@ -1,0 +1,10 @@
+package login;
+
+import InputAndOutput.SystemInOut;
+import UserFunctions.User;
+
+import java.io.IOException;
+
+public interface SignInGatewayInterface {
+    public User allowSignIn(String username, SystemInOut inOut) throws IOException, ClassNotFoundException;
+}
