+++ src/Main/login/SignUpGatewayInterface.java
@@ -1,0 +1,9 @@
+package login;
+
+import InputAndOutput.SystemInOut;
+
+import java.io.IOException;
+
+public interface SignUpGatewayInterface {
+    public void allowSignUp(String username) throws IOException, ClassNotFoundException;
+}
